lint and other nit fixes.
